nexus/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.12
cryptography==43.0.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
firebase-auth==3.0.28
gcloud==0.17.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.6
jws==0.1.3
msgpack==1.0.8
mysqlclient==2.2.4
numpy==2.0.1
oauth2client==3.0.0
packaging==24.1
pipdeptree==2.23.1
proto-plus==1.24.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.10.1
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-jwt==2.0.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==0.10.1
rest-framework-simplejwt==0.0.2
rsa==4.9
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
sseclient==0.0.27
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
wheel==0.43.0
